Compare grid search individuals only for the same gene
Before we would track the best so far and only update if it was better.
Because of outliers we would start making lots of updates and after a
few iterations barely make any. Now for each gene being changed we pick
the best local score there. This guarantees that we are always making
progress for each gene.